requirements.txt
gunicorn==21.2.0
flask==2.3.3
supabase==1.0.3
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
postgrest-py>=0.10.0,<0.11.0
gotrue>=2.1.0,<3.0.0